Basic scraper and html parsing
The scraper and parser is based on Cheerio, grabbing the content from the page and then
attempting to grab the elements needed from the HTML that's returned. Likely regex cou$
be better, but this is doing what I need for now.

html.js - has the basic code for grabbing links and images
server.js - will grab a single page and call the methods from html.js. Right now, it
also displays a web page of the informatio that is grabbed from the page